# collection.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "postman_collection",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "paths": {
...
    "/api/notifications": {
      "get": {
        "summary": "Get Notifications",
        "description": "The API endpoint retrieves notifications for a specific user identified by the userId parameter.\n\n**Generated using Postbot AI\n\n### Response\n\nThe response returns a JSON object with a key \"notification\" containing an array of notification objects. Each notification object has the following properties:\n\n- `id` (number): The unique identifier for the notification.\n    \n- `userId` (number): The identifier of the user to whom the notification belongs.\n    \n- `message` (string): The content of the notification message.\n    \n- `isRead` (boolean): Indicates whether the notification has been read or not.\n    \n- `createdAt` (string): The timestamp indicating when the notification was created.",
        "operationId": "getNotifications",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {}
        ]
      },
      "patch": {
        "summary": "Set Notification as Read",
        "description": "### Update Notification\n\n**Generated using Postbot AI\n\nThis endpoint is used to update a specific notification by its ID.\n\n#### Request\n\n- Method: PATCH\n    \n- URL: `http://localhost:3000/api/notifications?notificationId=1`\n    \n- Body: N/A\n    \n- `message` property.",
        "operationId": "setNotificationAsRead",
        "parameters": [
          {
            "name": "notificationId",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {}
        ]
      }
    },
...
    "/api/user/update": {
      "patch": {
        "summary": "Edit User",
        "description": "### Update User Details\n\n**Generated using Postbot AI\n\nThis endpoint is used to update the details of a specific user.\n\n#### Request\n\n- Method: PATCH\n    \n- URL: `http://localhost:3000/api/user/update?userId=1`\n    \n- Headers:\n    \n    - Content-Type: application/json",
        "operationId": "editUser",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phoneNumber": {
                    "type": "string",
                    "example": "56789109999128"
                  }
                }
              },
              "examples": {
                "Edit User": {
                  "value": {
                    "phoneNumber": "56789109999128"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/hotels/index": {
      "get": {
        "summary": "Get/List Hotels",
        "description": "This endpoint makes an HTTP GET request to retrieve a list of hotels based on the provided parameters. The request includes query parameters such as city, check-in date, check-out date, hotel name, star rating, minimum price, maximum price, and amenities.\n\n**Genereated using PostbotAI\n\n### Request Parameters\n\n- `city` (string): The city for which hotels are being searched.\n    \n- `checkInDate` (string): The check-in date for the hotel stay.\n    \n- `checkOutDate` (string): The check-out date for the hotel stay.\n    \n- `name` (string): The name of the hotel being searched.\n    \n- `starRating` (number): The star rating of the hotel.\n    \n- `minPrice` (number): The minimum price range for the hotel stay.\n    \n- `maxPrice` (number): The maximum price range for the hotel stay.\n    \n- `amenities` (string): The amenities desired, separated by commas.\n    \n\n### Response\n\nThe response will include a list of hotels that match the provided criteria. Each hotel object may contain details such as name, address, star rating, price, and amenities.",
        "operationId": "getListHotels",
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "summary": "Create Hotel",
        "description": "### POST /api/hotels/index\n\n**Genereated using PostbotAI\n\nThis endpoint is used to create a new hotel entry in the database.\n\n#### Request Body\n\n- name (string, required): The name of the hotel.\n    \n- address (string, required): The address of the hotel.\n    \n- location (string, required): The location of the hotel.\n    \n- starRating (number, required): The star rating of the hotel.\n    \n- images (array of strings, required): The images of the hotel.",
        "operationId": "createHotel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "string",
                    "example": "123 Main Street"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "image1.jpg"
                    },
                    "example": [
                      "image1.jpg",
                      "image2.png"
                    ]
                  },
                  "location": {
                    "type": "string",
                    "example": "Toronto"
                  },
                  "name": {
                    "type": "string",
                    "example": "Grandddd Hotel"
                  },
                  "starRating": {
                    "type": "number",
                    "example": 4.5
                  }
                }
              },
              "examples": {
                "Create Hotel": {
                  "value": {
                    "address": "123 Main Street",
                    "images": [
                      "image1.jpg",
                      "image2.png"
                    ],
                    "location": "Toronto",
                    "name": "Grandddd Hotel",
                    "starRating": 4.5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}